Restructure for PPTX, PDF service
If this only uses one service, the plugin adding headless Chrome is
added to all functions which increases their size by about 50MB. The PDF
function is the only one that uses Chromeless, so it's split out into
its own service. PPTX is working with a modification to use promises and
workarounds for quirks in the library. In the future, there will need to
be a strategy for deploying all of these behind a single API Gateway
endpoint, which is currently issue 3048 on the main serverless repo.